The peculiar thing about this plugin is that it doesn't work on the visual selection; it pastes above and then removes the selected line
That's not an operation, so 1v won't restore that, but the previous visual selection. In effect, an overridden repeat count gets lost. In case of a repeat with an overridden count, switch the repeat mapping from the ...Visual to the ...Line variant, so that future repeats use the new [count] lines instead of the original number of selected lines.